+++ testdata/refresh.sh
@@ -2,3 +2,5 @@
 
 curl -O https://raw.githubusercontent.com/web-platform-tests/wpt/master/url/resources/setters_tests.json
 curl -O https://raw.githubusercontent.com/web-platform-tests/wpt/master/url/resources/urltestdata.json
+curl -O https://raw.githubusercontent.com/web-platform-tests/wpt/master/url/resources/toascii.json
+curl -O https://raw.githubusercontent.com/web-platform-tests/wpt/master/url/resources/IdnaTestV2.json
